Clarify that "CWD" may by difffent from .eslintrc
We ran into this and spent some time debugging.  In particular the VS Code ESLint plug-in falls down if there are tsconfig.json files down the directory tree referenced by parallel .eslint files, and the `project` value is just "tsconfig.json" in each case.  That will always point to the *root* `tsconfig.json` which may be pretty unexpected.  Every .eslintrc file has to include the full path to the tsconfig.json, even if it's in the same directory.

Issue for reference: https://github.com/typescript-eslint/typescript-eslint/issues/2434